# colblk: implement HideObsoletePoints in DataBlockIter