@jbowens 2024-09-30
```
--- FAIL: TestLint (42.48s)
    --- FAIL: TestLint/TestGCAssert (10.61s)
        lint_test.go:129: See /tmp/gcassert-1156708372.log for full output.
        lint_test.go:132: 
            sstable/colblk/data_block.go:802:	i.setNextObsoletePoint(): call was not inlined
        lint_test.go:132: 
            sstable/colblk/data_block.go:831:	i.setNextObsoletePoint(): call was not inlined
        lint_test.go:132: 
            sstable/colblk/data_block.go:846:	i.setNextObsoletePoint(): call was not inlined
        lint_test.go:132: 
            sstable/colblk/data_block.go:931:	i.setNextObsoletePoint(): call was not inlined
        lint_test.go:132: 
            sstable/colblk/data_block.go:979:	i.setNextObsoletePoint(): call was not inlined
```

:'(