@RaduBerinde 2024-10-01
So strange, `SeekBitGE` is not inlined, so how can just an if and a function call be too much to inline? Full logs show `setNextObsoletePoint: function too complex: cost 81 exceeds budget 80`