@RaduBerinde 2024-10-01
Ok, moved the check to `SeekBitGE` and verified that it didn't affect the normal benchmarks:
```
name                                                                                old time/op    new time/op    delta
CockroachDataBlockIterShort/AlphaLen=8,Prefix=8,Shared=4,ValueLen=8/Next-10           12.1ns ± 0%    11.9ns ± 1%  -1.55%  (p=0.000 n=8+8)
CockroachDataBlockIterShort/AlphaLen=8,Prefix=8,Shared=4,ValueLen=8/SeekGE-10          113ns ± 0%     113ns ± 0%    ~     (p=0.631 n=8+8)
CockroachDataBlockIterShort/AlphaLen=8,Prefix=128,Shared=64,ValueLen=128/Next-10      10.1ns ± 0%    10.2ns ± 0%  +0.75%  (p=0.000 n=8+7)
CockroachDataBlockIterShort/AlphaLen=8,Prefix=128,Shared=64,ValueLen=128/SeekGE-10    89.6ns ± 0%    91.0ns ± 1%  +1.58%  (p=0.000 n=7+8)
```